PR kommentarer fikset
